Remove sidebar and header controls from auth routes
