+++ apps/web/src/app/(auth-pages)/layout.tsx
@@ -1,7 +1,6 @@
 import { Inter, Alfa_Slab_One } from "next/font/google";
 import { AuthLayoutContent } from "@/components/desktop/layout/auth/auth";
 import { cn } from "@/lib/utils";
-import "@/styles/globals.css";
 
 const inter = Inter({ subsets: ["latin"], variable: "--font-sans" });
 const alfaSlabOne = Alfa_Slab_One({
@@ -15,16 +14,5 @@ export default function AuthLayout({
 }: {
   children: React.ReactNode;
 }) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <head>
-        <link rel="icon" href="/rc-logo-b.png" sizes="any" />
-      </head>
-      <body
-        className={cn(inter.className, alfaSlabOne.variable, "overflow-hidden")}
-      >
-        <AuthLayoutContent>{children}</AuthLayoutContent>
-      </body>
-    </html>
-  );
+  return <AuthLayoutContent>{children}</AuthLayoutContent>;
 }
