+++ apps/web/src/app/(auth-pages)/sign-in/page.tsx
@@ -8,40 +8,55 @@ import { Input } from "@/components/core/ui/input";
 import { Label } from "@/components/core/ui/label";
 import Link from "next/link";
 
+// Debug component to show layout detection
+function LayoutDebug() {
+  return (
+    <div className="fixed top-0 left-0 m-1 p-2 text-xs bg-black/80 text-white rounded z-50">
+      Auth Page - Should NOT show sidebar
+    </div>
+  );
+}
+
 export default async function Login(props: { searchParams: Promise<Message> }) {
   const searchParams = await props.searchParams;
   return (
-    <form className="flex-1 flex flex-col min-w-64">
-      <h1 className="text-2xl font-medium">Sign in</h1>
-      <p className="text-sm text-foreground">
-        Don&apos;t have an account?{" "}
-        <Link className="text-foreground font-medium underline" href="/sign-up">
-          Sign up
-        </Link>
-      </p>
-      <div className="flex flex-col gap-2 [&>input]:mb-3 mt-8">
-        <Label htmlFor="email">Email</Label>
-        <Input name="email" placeholder="you@example.com" required />
-        <div className="flex justify-between items-center">
-          <Label htmlFor="password">Password</Label>
+    <>
+      {/* <LayoutDebug /> */}
+      <form className="flex-1 flex flex-col min-w-64">
+        <h1 className="text-2xl font-medium">Sign in</h1>
+        <p className="text-sm text-foreground">
+          Don&apos;t have an account?{" "}
           <Link
-            className="text-xs text-foreground underline"
-            href="/forgot-password"
+            className="text-foreground font-medium underline"
+            href="/sign-up"
           >
-            Forgot Password?
+            Sign up
           </Link>
+        </p>
+        <div className="flex flex-col gap-2 [&>input]:mb-3 mt-8">
+          <Label htmlFor="email">Email</Label>
+          <Input name="email" placeholder="you@example.com" required />
+          <div className="flex justify-between items-center">
+            <Label htmlFor="password">Password</Label>
+            <Link
+              className="text-xs text-foreground underline"
+              href="/forgot-password"
+            >
+              Forgot Password?
+            </Link>
+          </div>
+          <Input
+            type="password"
+            name="password"
+            placeholder="Your password"
+            required
+          />
+          <SubmitButton pendingText="Signing In..." formAction={signInAction}>
+            Sign in
+          </SubmitButton>
+          <FormMessage message={searchParams} />
         </div>
-        <Input
-          type="password"
-          name="password"
-          placeholder="Your password"
-          required
-        />
-        <SubmitButton pendingText="Signing In..." formAction={signInAction}>
-          Sign in
-        </SubmitButton>
-        <FormMessage message={searchParams} />
-      </div>
-    </form>
+      </form>
+    </>
   );
 }
