+++ apps/web/src/app/(routes)/layout.tsx
@@ -1,5 +1,4 @@
 import type { Metadata } from "next";
-
 import "@/styles/globals.css";
 
 export const metadata: Metadata = {
@@ -15,5 +14,6 @@ export default function RoutesLayout({
 }: {
   children: React.ReactNode;
 }) {
+  // We don't need to wrap with DashboardLayout here as that's handled in the root layout
   return <>{children}</>;
 }
