+++ apps/web/src/middleware.ts
@@ -7,6 +7,7 @@ const publicRoutes = [
   "/sign-in",
   "/sign-up",
   "/forgot-password",
+  "/reset-password",
   "/auth/callback",
 ];
 
@@ -16,12 +17,30 @@ type UpdateSessionResult = {
   session: Session | null;
 };
 
+  /**
+   * This middleware handles session authentication and redirects.
+   *
+   * The following logic applies:
+   * 1. If on an auth route and logged in, redirect to dashboard.
+   * 2. If on a protected route and not logged in, redirect to sign-in.
+   * 3. Otherwise, return the response.
+   *
+   * @param request - The Next.js request object.
+   * @returns The response to return to the client.
+   */
 export async function middleware(request: NextRequest) {
   // Get the response and session from updateSession
   const result = (await updateSession(request)) as UpdateSessionResult;
   const { response, session } = result;
 
   const pathname = request.nextUrl.pathname;
+  
+  // Set the pathname in response headers for layouts to access
+  response.headers.set("x-pathname", pathname);
+  
+  // Add debug header
+  response.headers.set("x-debug-is-auth-route", 
+    publicRoutes.some(route => pathname === route || pathname.startsWith(route)).toString());
 
   // Check if the current route is an auth route
   const isAuthRoute = publicRoutes.some(
